+++ src/main/java/at/hannibal2/skyhanni/events/DungeonEnterEvent.kt
@@ -1,0 +1,3 @@
+package at.hannibal2.skyhanni.events
+
+class DungeonEnterEvent(dungeonFloor: String): LorenzEvent()